Ensure using float64 and numpy arrays everywhere

+++ src/optilab/functions/surrogate/knn_surrogate_objective_function.py
@@ -47,11 +47,13 @@ class KNNSurrogateObjectiveFunction(SurrogateObjectiveFunction):
         super().train(train_set)
 
         x_train, y_train = self.train_set.pairs()
-        x_train = np.array(x_train, dtype=np.float32)
-        y_train = np.array(y_train, dtype=np.float32)
+        x_train = np.array(x_train, dtype=np.float64)
+        y_train = np.array(y_train, dtype=np.float64)
 
         self.faiss_index = faiss.IndexFlatL2(x_train.shape[1])
-        self.faiss_index.add(x_train)  # pylint: disable=no-value-for-parameter
+        self.faiss_index.add(  # pylint: disable=no-value-for-parameter
+            x_train.astype(np.float32)
+        )
         self.y_train = y_train
 
     def __call__(self, point: Point) -> Point:
@@ -69,13 +71,14 @@ class KNNSurrogateObjectiveFunction(SurrogateObjectiveFunction):
         if len(self.train_set) < self.num_neighbors:
             raise ValueError("Train set length is below number of neighbors.")
 
-        x_query = np.array([point.x], dtype=np.float32)
+        x_query = np.array([point.x], dtype=np.float64)
         distances, indices = (
             self.faiss_index.search(  # pylint: disable=no-value-for-parameter
-                x_query,
+                x_query.astype(np.float32),
                 self.num_neighbors,
             )
         )
+        distances = distances.astype(np.float64)
 
         weights = 1 / (np.sqrt(distances) + 1e-8)  # avoid division by zero
         y_pred = np.sum(self.y_train[indices] * weights, axis=1)[0] / weights.sum()
